moving files into logical folders
